backend/requirements.txt
fastapi
uvicorn
pandas
numpy
pydantic
python-multipart
PyMuPDF
scikit-learn
sentence-transformers>=3.0.0
torch>=2.7.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
pymupdf